增加一些方法

+++ appLib2.0/src/main/java/com/xiay/applib/view/recyclerview/recyclerviewflexibledivider/GridSpacingItemDecoration.java
@@ -15,7 +15,7 @@ public class GridSpacingItemDecoration extends RecyclerView.ItemDecoration {
     private int spacing;
     private boolean includeEdge;
 
-    public GridSpacingItemDecoration(int spanCount, int spacing, boolean includeEdge) {
+    public GridSpacingItemDecoration(int spanCount, float spacing, boolean includeEdge) {
         this.spanCount = spanCount;
         this.spacing =  ViewUtil.scaleValue(spacing);
         this.includeEdge = includeEdge;
